requirements/dev.txt
-r prod.txt
-r test.txt
black
coverage==5.1
coveralls==3.3.1
mypy
pylint
pylint-pytest
py-githooks
rope
twine
wheel
